Fixed hierarchy of proper server and component clients in login and sign up page.
