+++ src/app/pages/moa-status/page.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 // Components
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
