+++ src/components/ui/card.jsx
@@ -1,3 +1,5 @@
+"use client"
+
 import * as React from "react";
 
 import { cn } from "@/lib/utils";
